Change semantic_taint_threshold from voxel count to ratio
A segment is tainted when taint_voxels / total_voxels > threshold,
which avoids small fragments with stray taint voxels being marked.

+++ waterz/__init__.py
@@ -32,7 +32,7 @@ def agglomerate(
         semantic_signal_ratio: float = 0.6,
 
         semantic_taint_labels: list[int] = [],
-        semantic_taint_threshold: int = 0,
+        semantic_taint_threshold: float = 0.0,
 
         size_heuristic_aff_threshold: float = 1.0,  # disabled
         size_heuristic_small_threshold: int = 1_000_000,
